+++ app/components/job-schedule/details/job-schedule-details.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from "@angular/core";
 
 import { BaseModule } from "@batch-flask/ui";
 import { commonModules } from "app/common";
+import { JobActionModule } from "app/components/job/action";
 import { JobDetailsModule } from "app/components/job/details";
 import { PoolDetailsModule } from "app/components/pool/details";
 import { JobScheduleAutoPoolComponent } from "./job-schedule-autopool.component";
@@ -22,6 +23,7 @@ const modules = [
     BaseModule,
     JobDetailsModule,
     PoolDetailsModule,
+    JobActionModule,
 ];
 
 @NgModule({
