+++ app/components/job/action/add-task/index.ts
@@ -1,0 +1,1 @@
+export * from "./add-task-button.component";
