+++ app/components/job/action/disable/disable-job-button.component.ts
@@ -1,0 +1,27 @@
+import { ChangeDetectionStrategy, Component, Input } from "@angular/core";
+import { BaseButton } from "@batch-flask/ui/buttons";
+import { Job, JobState } from "app/models/job";
+
+@Component({
+    selector: "bl-disable-button",
+    template: `
+        <bl-button color="light" [action]="action" *ngIf="visible" icon="fa fa-pause"
+            [disabled]="!enabled" title="Disable" permission="write">
+        </bl-button>
+    `,
+    changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class DisableJobButtonComponent extends BaseButton {
+    @Input() public job: Job;
+
+    public get enabled() {
+        return this.job
+            && this.job.state === JobState.active;
+    }
+
+    public get visible() {
+        return this.job
+            && this.job.state !== JobState.disabling
+            && this.job.state !== JobState.disabled;
+    }
+}
