+++ app/components/job/action/terminate/index.ts
@@ -1,0 +1,2 @@
+export * from "./terminate-job-dialog.component";
+export * from "./terminate-button.component";
