+++ src/@batch-flask/ui/table/table.component.spec.ts
@@ -8,14 +8,13 @@ import { ContextMenuService } from "@batch-flask/ui/context-menu";
 import {
     TableCellComponent, TableColumnComponent, TableComponent, TableHeadComponent, TableRowComponent,
 } from "@batch-flask/ui/table";
-import { VmSize } from "app/models";
 import { click } from "test/utils/helpers";
 import { virtualScrollMockComponents } from "test/utils/mocks/components";
 
-const sizeA = new VmSize({ name: "Size A", numberOfCores: 1, resourceDiskSizeInMB: 1000 } as any);
-const sizeB = new VmSize({ name: "Size B", numberOfCores: 8, resourceDiskSizeInMB: 2000 } as any);
-const sizeC = new VmSize({ name: "Size C", numberOfCores: 4, resourceDiskSizeInMB: 80000 } as any);
-const sizeD = new VmSize({ name: "Size D", numberOfCores: 2, resourceDiskSizeInMB: 4000 } as any);
+const sizeA = { name: "Size A", numberOfCores: 1, resourceDiskSizeInMB: 1000 };
+const sizeB = { name: "Size B", numberOfCores: 8, resourceDiskSizeInMB: 2000 };
+const sizeC = { name: "Size C", numberOfCores: 4, resourceDiskSizeInMB: 80000 };
+const sizeD = { name: "Size D", numberOfCores: 2, resourceDiskSizeInMB: 4000 };
 
 // tslint:disable:trackBy-function
 @Component({
